pyproject.toml
[tool.poetry]
name = "market-data-vwap-snapper"
version = "0.0.1"
description = "Market Data via websocket and analytics"
authors = [
    "Edoardo Busetti <edoardo@busetti.it>",
]

[tool.poetry.dependencies]
python = ">=3.11,<3.11.13"
requests = "^2.28"
numpy = ">=1.22.2"
sqlalchemy = "^2.0.23"
bintrees = "^2.2.0"
python-dateutil = "^2.8.2"
websockets = "^12.0"
cachetools = "^5.3.2"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
psycopg2 = "^2.9.9"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
line_length = 99